# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "V-LAB PROJECT DOCUMENTATION",
        "contact": {
            "email": "man2@cin.ufpe.br"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/transaction/{id_user}": {
            "get": {
                "tags": [
                    "Transação"
                ],
                "summary": "Listar todas as Transações do Usuário Especificado",
                "description": "Usuário pode utilizar para visualizar todas as suas transações.",
                "operationId": "transactionIndex",
                "parameters": [
                    {
                        "name": "id_user",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operação Bem Seucedida"
                    },
                    "404": {
                        "description": "Nenhum Usuário Encontrado"
                    }
                }
            },
            "post": {
                "tags": [
                    "Transação"
                ],
                "summary": "Cadastrar Transação Para o Usuário Especificado",
                "description": "Usuário pode utilizar para cadastrar uma nova transação na sua conta.",
                "operationId": "transactionStore",
                "parameters": [
                    {
                        "name": "id_user",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Dados da Categoria",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "value",
                                    "type",
                                    "id_category"
                                ],
                                "properties": {
                                    "value": {
                                        "type": "numeric",
                                        "example": ""
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "id_category": {
                                        "type": "integer",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Input Inválido"
                    },
                    "201": {
                        "description": "Transação Adicionada no Banco de Dados"
                    },
                    "500": {
                        "description": "Algo Errado no Insert"
                    }
                }
            }
        },
        "/api/transaction/{id_user}/delete/{id_category}": {
            "delete": {
                "tags": [
                    "Transação"
                ],
                "summary": "Deletar a Transação Especificada do Usuário Especificado",
                "description": "Usuário pode utilizar para deletar uma transação da sua conta.",
                "operationId": "transactionDestroy",
                "parameters": [
                    {
                        "name": "id_user",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "id_category",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transação Excluída"
                    },
                    "404": {
                        "description": "Transação não Encontrado"
                    }
                }
            }
        },
        "/api/transaction": {
            "get": {
                "tags": [
                    "Transação"
                ],
                "summary": "Listar todas as Transações Existentes",
                "description": "Desenvolvedor pode utilizar para visualizar todas as transações existentes no banco de dados.",
                "operationId": "transactionIndexAdmin",
                "responses": {
                    "200": {
                        "description": "Operação Bem Seucedida"
                    },
                    "404": {
                        "description": "Nenhuma Transação Encontrada"
                    }
                }
            }
        },
...
    }
}